Extract test filter logic into hook
