+++ src/components/HistoryTable/index.tsx
@@ -5,6 +5,7 @@ import * as context from "./HistoryTableContext";
 import { HistoryTableTestSearch } from "./HistoryTableTestSearch";
 import Row from "./Row";
 import * as types from "./types";
+import useTestFilters from "./useTestFilters";
 import useTestResults from "./useTestResults";
 
 export {
@@ -15,5 +16,6 @@ export {
   types,
   useTestResults,
   HistoryTableTestSearch,
+  useTestFilters,
 };
 export default HistoryTable;
