+++ src/pages/VariantHistory.tsx
@@ -8,6 +8,7 @@ import HistoryTable, {
   context,
   ColumnPaginationButtons,
   HistoryTableTestSearch,
+  useTestFilters,
 } from "components/HistoryTable";
 import { PageWrapper } from "components/styles";
 import { size } from "constants/tokens";
@@ -91,7 +92,8 @@ export const VariantHistoryContents: React.FC = () => {
     },
   });
 
-  const { setHistoryTableFilters, addColumns } = useHistoryTable();
+  const { addColumns } = useHistoryTable();
+  useTestFilters();
 
   const { taskNamesForBuildVariant } = columnData || {};
   const { mainlineCommits } = data || {};
@@ -99,20 +101,7 @@ export const VariantHistoryContents: React.FC = () => {
   const selectedTaskNames = useMemo(() => toArray(queryParams.tasks), [
     queryParams.tasks,
   ]);
-  useEffect(() => {
-    const failingTests = toArray(queryParams[TestStatus.Failed]);
-    const passingTests = toArray(queryParams[TestStatus.Passed]);
-    const failingTestFilters = failingTests.map((test) => ({
-      testName: test,
-      testStatus: TestStatus.Failed,
-    }));
-    const passingTestFilters = passingTests.map((test) => ({
-      testName: test,
-      testStatus: TestStatus.Passed,
-    }));
-    setHistoryTableFilters([...failingTestFilters, ...passingTestFilters]);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [queryParams]);
+
   const queryParamsToDisplay = new Set([
     TestStatus.Failed,
     TestStatus.Passed,
